requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
bcrypt==4.3.0
blinker==1.9.0
build==1.2.2.post1
CacheControl==0.14.2
cachetools==5.5.2
captcha==0.7.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cleo==2.1.0
click==8.1.8
colorama==0.4.6
crashtest==0.4.1
cryptography==44.0.2
distlib==0.3.9
distro==1.9.0
dulwich==0.22.8
extra-streamlit-components==0.1.71
fastjsonschema==2.21.1
filelock==3.18.0
findpython==0.6.3
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.1.1
groq==0.20.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
installer==0.7.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keyring==25.6.0
langchain==0.3.21
langchain-core==0.3.46
langchain-groq==0.3.0
langchain-text-splitters==0.3.7
langsmith==0.3.18
MarkupSafe==3.0.2
more-itertools==10.6.0
msgpack==1.1.0
narwhals==1.31.0
numpy==2.2.4
orjson==3.10.15
packaging==24.2
pandas==2.2.3
pbs-installer==2025.3.17
pillow==11.1.0
pkginfo==1.12.1.2
platformdirs==4.3.7
poetry-core==2.1.1
protobuf==5.29.4
psycopg2==2.9.10
pyarrow==19.0.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
pydotenv==0.0.7
PyJWT==2.10.1
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
pywin32-ctypes==0.2.3
PyYAML==6.0.2
RapidFuzz==3.12.2
referencing==0.36.2
requests==2.32.3
requests-toolbelt==1.0.0
rpds-py==0.23.1
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.39
streamlit==1.43.2
streamlit-authenticator==0.4.2
tenacity==9.0.0
toml==0.10.2
tomlkit==0.13.2
tornado==6.4.2
trove-classifiers==2025.3.19.19
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
virtualenv==20.29.3
watchdog==6.0.0
zstandard==0.23.0
